robotclient/requirements.txt
pyserial==3.4
PyYAML==3.13
robotremoteserver==1.1
